Move biomes to their own file
