+++ rgen/rgen-base/src/lib.rs
@@ -1,3 +1,4 @@
+mod biome;
 mod block;
 mod chunk;
 mod filter;
@@ -5,9 +6,8 @@ mod iter;
 mod pos;
 mod prop;
 
-pub use block::{
-  Biome, BiomeId, BlockData, BlockId, BlockInfo, BlockKind, BlockState, StateId, StateOrProps,
-};
+pub use biome::{Biome, BiomeId};
+pub use block::{BlockData, BlockId, BlockInfo, BlockKind, BlockState, StateId, StateOrProps};
 pub use chunk::Chunk;
 pub use filter::BlockFilter;
 pub use iter::{BlocksIterExclusive, BlocksIterInclusive};
